+++ scripts/add_course_images.js
@@ -1,7 +1,7 @@
 // ==UserScript==
 // @name         UoM Blackboard: Add course images
 // @namespace    http://tampermonkey.net/
-// @version      20240119.00.00
+// @version      20240119.21.00
 // @description  An optional accompanying script for https://github.com/adil192/BlackboardTheme, which adds course images to the Blackboard homepage.
 // @author       adil192
 // @match        https://online.manchester.ac.uk/webapps/portal/*
@@ -14,45 +14,61 @@
 // @ts-check
 
 /**
- * Returns a module image URL, assuming it's in standard format.
- * @param {string} moduleCode
- * @param {string} extension The file extension of the image, defaults to "jpg".
- * @returns {string}
+ * A string that is used as a placeholder in the module images database
+ * when no Pixabay images are found for a module.
+ * 
+ * This is used to prevent the script from trying to find images for the
+ * same module again and again, using up the Pixabay API quota.
  */
-function imageFromModuleCode(moduleCode, extension="jpg") {
-    return `https://raw.githubusercontent.com/adil192/BlackboardTheme/main/assets/subjects/${moduleCode}/${moduleCode}.${extension}`;
-}
+const NORESULTS = "NORESULTS";
 
 /**
- * The known modules and their images.
- * In the future, I want to replace this with the Pixabay API
- * so that we can get images for any module.
- * @type {[string, string][]}
+ * Known modules whose images aren't from Pixabay.
+ * @type {Record<string, string>}
  */
-const moduleImages = [
-    ["ARDSE001",  imageFromModuleCode("ARDSE001")],
-    ["COMP",      imageFromModuleCode("COMP")],
-    ["COMP30040", imageFromModuleCode("COMP30040")],
-    ["COMP33511", imageFromModuleCode("COMP33511")],
-    ["COMP38311", imageFromModuleCode("COMP38311")],
-    ["MATH",      imageFromModuleCode("MATH")],
-    ["MATH31051", imageFromModuleCode("MATH31051", "png")],
-    ["MATH32031", imageFromModuleCode("MATH32031")],
-    ["MATH32091", imageFromModuleCode("MATH32091", "png")],
-];
+const knownModuleImages = {
+    "ARDSE001" : "https://raw.githubusercontent.com/adil192/BlackboardTheme/main/assets/subjects/ARDSE001/ARDSE001.jpg",
+    "COMP30040": "https://raw.githubusercontent.com/adil192/BlackboardTheme/main/assets/subjects/COMP30040/COMP30040.jpg",
+    "COMP33511": "https://raw.githubusercontent.com/adil192/BlackboardTheme/main/assets/subjects/COMP33511/COMP33511.jpg",
+    "COMP38311": "https://raw.githubusercontent.com/adil192/BlackboardTheme/main/assets/subjects/COMP38311/COMP38311.jpg",
+    "MATH31051": "https://raw.githubusercontent.com/adil192/BlackboardTheme/main/assets/subjects/MATH31051/MATH31051.png",
+    "MATH32031": "https://raw.githubusercontent.com/adil192/BlackboardTheme/main/assets/subjects/MATH32031/MATH32031.jpg",
+    "MATH32091": "https://raw.githubusercontent.com/adil192/BlackboardTheme/main/assets/subjects/MATH32091/MATH32091.png",
+};
 
 /**
- * Returns the module image for the given module name, or null if it doesn't exist.
- * @param {string} moduleName
- * @returns {string | null}
+ * These are used as fallbacks if the module isn't in `knownModuleImages`
+ * and no Pixabay images are found.
+ * @type {Record<string, string>}
  */
-function findModuleImage(moduleName) {
-    let bestImage = null;
-    for (const [moduleCode, moduleImage] of moduleImages) {
-        if (!moduleName.startsWith(moduleCode)) continue;
-        bestImage = moduleImage;
-    }
-    return bestImage;
+const fallbackModuleImages = {
+    "COMP": "https://raw.githubusercontent.com/adil192/BlackboardTheme/main/assets/subjects/COMP/COMP.jpg",
+    "MATH": "https://raw.githubusercontent.com/adil192/BlackboardTheme/main/assets/subjects/MATH/MATH.jpg",
+};
+
+/**
+ * The IndexDB database with the cached module images.
+ * @type {IDBDatabase | null}
+ */
+let moduleImagesDB = null;
+
+/**
+ * Opens the IndexDB database with the cached module images.
+ * @returns {Promise<void>}
+ */
+function openDb() {
+    return new Promise((resolve, reject) => {
+        const request = indexedDB.open("moduleImages", 1);
+        request.onerror = reject;
+        request.onsuccess = () => {
+            moduleImagesDB = request.result;
+            resolve();
+        };
+        request.onupgradeneeded = () => {
+            const db = request.result;
+            db.createObjectStore("moduleImages");
+        };
+    });
 }
 
 /** @type {HTMLDivElement[]} */
@@ -90,6 +106,132 @@ async function findCoursesDivs() {
 }
 
 /**
+ * Returns the module image for the given module name, or null if it doesn't exist.
+ * @param {string} moduleName
+ * @returns {Promise<string | null>}
+ */
+async function findModuleImage(moduleName) {
+    if (!moduleImagesDB) {
+        console.error('moduleImagesDB not open');
+        return null;
+    }
+
+    for (const moduleCode in knownModuleImages) {
+        if (!moduleName.startsWith(moduleCode)) continue;
+        return knownModuleImages[moduleCode];
+    }
+
+    const transaction = moduleImagesDB.transaction("moduleImages", "readonly");
+    const objectStore = transaction.objectStore("moduleImages");
+    const cachedImage = await new Promise((resolve) => {
+        const request = objectStore.get(moduleName);
+        request.onsuccess = () => resolve(request.result);
+        request.onerror = () => resolve(null);
+    });
+
+    if (cachedImage) {
+        if (cachedImage === NORESULTS) return null;
+        else return cachedImage;
+    } else {
+        return await findAndCacheModuleImage(moduleName);
+    }
+}
+
+/**
+ * Finds a module image from Pixabay for the given module name,
+ * and caches it in the module images database.
+ * 
+ * 
+ * @param {string} moduleName
+ * @returns {Promise<string | null>}
+ */
+async function findAndCacheModuleImage(moduleName) {
+    const apiKey = "25286000-bf7eb8ff8e6d2d1630cf59fae";
+    const keywords = extractKeywords(moduleName);
+    const url = `https://pixabay.com/api/?key=${apiKey}&q=${encodeURIComponent(keywords)}&image_type=photo&per_page=3`;
+    console.log(`Searching Pixabay for ${moduleName}: ${url}`);
+    const response = await fetch(url);
+    const json = await response.json();
+    const hits = json.hits;
+    if (hits.length === 0) {
+        console.error(`No Pixabay images found for ${moduleName}`);
+
+        // Don't try again for this module.
+        if (moduleImagesDB) {
+            const transaction = moduleImagesDB.transaction("moduleImages", "readwrite");
+            const objectStore = transaction.objectStore("moduleImages");
+            objectStore.put(NORESULTS, moduleName);
+        }
+        return null;
+    }
+    const moduleImage = hits[0].webformatURL;
+    const dataUrl = await urlToDataUrl(moduleImage);
+
+    if (moduleImagesDB) {
+        const transaction = moduleImagesDB.transaction("moduleImages", "readwrite");
+        const objectStore = transaction.objectStore("moduleImages");
+        objectStore.put(dataUrl, moduleName);
+    }
+    return dataUrl;
+}
+
+/**
+ * Converts an image URL to a data URL.
+ * @param {string} url
+ * @returns {Promise<string>}
+ */
+function urlToDataUrl(url) {
+    return new Promise((resolve) => {
+        const canvas = document.createElement("canvas");
+        const context = canvas.getContext("2d");
+        if (!context) throw new Error("Failed to get canvas context");
+        const image = new Image();
+        image.addEventListener("load", () => {
+            canvas.width = image.width;
+            canvas.height = image.height;
+            context.drawImage(image, 0, 0);
+            const dataUrl = canvas.toDataURL("image/webp");
+            resolve(dataUrl);
+        });
+        image.setAttribute("crossorigin", "anonymous");
+        image.src = url;
+    });
+}
+
+/**
+ * Extracts the keywords from a module name.
+ *
+ * E.g. "COMP30040&30030 Third Year Project Laboratory 2023-24 Full Year"
+ * becomes "Project Laboratory".
+ * 
+ * E.g. "COMP33511 User Experience 2023-24 1st Semester"
+ * becomes "User Experience".
+ *
+ * @param {string} moduleName
+ * @returns {string}
+ */
+function extractKeywords(moduleName) {
+    let words = moduleName.split(" ");
+    
+    // remove any words that contain a number
+    words = words.filter((word) => !/\d/.test(word));
+
+    // remove banned words
+    const bannedWords = [
+        "Semester", "Full", "Year",
+        "and", "or", "the", "of", "in", "to", "for", "with", "on", "at", "from",
+        "First", "Second", "Third", "Fourth", "Fifth",
+        "I", "II", "III", "IV", "V",
+    ];
+    words = words.filter((word) => !bannedWords.includes(word));
+
+    // limit to the first 4 words
+    words = words.slice(0, 4);
+
+    return words.join(" ");
+}
+
+/**
  * Sets the `data-module-code` attribute of each course to the course code.
  */
 function labelCourses() {
@@ -106,10 +248,13 @@ function labelCourses() {
             course.dataset.moduleCode = moduleName;
 
             // Set `--bg-url` to the module image if it exists.
-            let moduleImage = findModuleImage(moduleName);
-            if (moduleImage) {
-                course.style.setProperty("--bg-url", `url("${moduleImage}")`);
+            if (fallbackModuleImages[moduleName]) {
+                course.style.setProperty("--bg-url", `url("${fallbackModuleImages[moduleName]}")`);
             }
+            findModuleImage(moduleName).then((moduleImage) => {
+                if (!moduleImage) return;
+                course.style.setProperty("--bg-url", `url("${moduleImage}")`);
+            });
         });
     });
 }
@@ -118,7 +263,11 @@ function labelCourses() {
     'use strict';
 
     console.log("UoM Blackboard: Add course images");
-    findCoursesDivs().then(() => {
+
+    Promise.all([
+        findCoursesDivs(),
+        openDb(),
+    ]).then(() => {
         labelCourses();
 
         coursesDivs.forEach((coursesDiv) => {